requirements.txt
alembic==1.11.1
amqp==5.1.1
astroid==2.15.5
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.6
Flask==2.3.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
platformdirs==3.5.1
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.7.0
pylint==2.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.15
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
